# pharma-platform/server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PharmaGuard API
  version: 1.0.0
  description: |
    Pharmaceutical Data Integrity & Compliance Platform.
    FDA 21 CFR Part 11 compliant electronic records, audit trails, 
    e-signatures, batch tracking, and document control APIs.
  contact:
    name: PharmaGuard Engineering
    email: api@pharmaguard.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development
  - url: https://api.pharmaguard.io/api/v1
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for watch agents and external integrations
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: "priya.sharma@pharmaguard.io" }
        password: { type: string, format: password, example: "PharmaGuard@2026" }
    LoginResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                id: { type: string, format: uuid }
                email: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                role: { type: string, enum: [admin, lab_director, qa_manager, qc_analyst, compliance_officer, lab_technician, viewer] }
                department: { type: string }
                orgId: { type: string, format: uuid }
            tokens:
              type: object
              properties:
                accessToken: { type: string }
                refreshToken: { type: string }
...
    AuditEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        userEmail: { type: string }
        userName: { type: string }
        userRole: { type: string }
        action: { type: string }
        category: 
          type: string
          enum: [data_capture, integrity, document_control, access_control, batch_tracking, signature, instrument, reporting, system]
        severity:
          type: string
          enum: [info, warning, critical]
        resourceType: { type: string }
        resourceId: { type: string, format: uuid }
        description: { type: string }
        entryHash: { type: string, description: "SHA-256 hash of entry for tamper evidence" }
        previousHash: { type: string, description: "Hash of previous entry in chain" }
        sequenceNumber: { type: integer }
        createdAt: { type: string, format: date-time }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { description: Invalid credentials }
        '423': { description: Account locked }
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: New tokens issued }
...
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 12 }
      responses:
        '200': { description: Password changed }

  # ─── File Uploads ──────────────────────────────────────
...
  /audit:
    get:
      tags: [Audit Trail]
      summary: Query audit trail
      description: Retrieve immutable audit entries with comprehensive filtering
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, maximum: 100 } }
        - { name: userId, in: query, schema: { type: string, format: uuid } }
        - { name: action, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: severity, in: query, schema: { type: string } }
        - { name: resourceType, in: query, schema: { type: string } }
        - { name: startDate, in: query, schema: { type: string, format: date-time } }
        - { name: endDate, in: query, schema: { type: string, format: date-time } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated audit trail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/AuditEntry' } }
                  summary: 
                    type: object
                    properties:
                      info: { type: integer }
                      warning: { type: integer }
                      critical: { type: integer }
                  pagination: { $ref: '#/components/schemas/Pagination' }
...
  /audit/resource/{type}/{resourceId}:
    get:
      tags: [Audit Trail]
      summary: Audit history for specific resource
      parameters:
        - { name: type, in: path, required: true, schema: { type: string } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Resource audit trail }
  /audit/stats/summary:
    get:
      tags: [Audit Trail]
      summary: Audit trail statistics
      parameters:
        - { name: days, in: query, schema: { type: integer, default: 30 } }
      responses:
        '200': { description: Audit statistics }

  # ─── Documents ─────────────────────────────────────────
...